Add some type file preview and more (#28)
Adds preview for scribus file .sla. fix #25
Adds test for gimp file .xfc. fix #26
Refactor openoffice and scribus preview builder.
Complete README.
Fix bug with exif tool. fix #21
Adds some PEP8.

+++ preview_generator/preview/builder/office__libreoffice.py
@@ -6,29 +6,24 @@ import os
 from subprocess import check_call
 from subprocess import DEVNULL
 from subprocess import STDOUT
-import time
 import typing
-
-from PyPDF2 import PdfFileReader
-from PyPDF2 import PdfFileWriter
 
 from preview_generator.exception import BuilderDependencyNotFound
 from preview_generator.exception import ExecutableNotFound
-from preview_generator.preview.generic_preview import PreviewBuilder
 from preview_generator.utils import check_executable_is_available
-from preview_generator.utils import ImgDims
-from preview_generator.preview.builder.image__wand import convert_pdf_to_jpeg
-from pathlib import Path
+from preview_generator.preview.builder.document_generic import DocumentPreviewBuilder
+from preview_generator.preview.builder.document_generic import create_flag_file
+from preview_generator.preview.builder.document_generic import write_file_content
 
 
-class OfficePreviewBuilderLibreoffice(PreviewBuilder):
+class OfficePreviewBuilderLibreoffice(DocumentPreviewBuilder):
     @classmethod
     def get_label(cls) -> str:
         return 'Documents - based on LibreOffice'
 
     @classmethod
     def get_supported_mimetypes(cls) -> typing.List[str]:
-        return LO_MIMETYPES.keys()  # type: typing.List[str]
+        return [k for k in typing.cast(str, LO_MIMETYPES.keys())]
 
     @classmethod
     def check_dependencies(cls) -> bool:
@@ -39,175 +34,17 @@ class OfficePreviewBuilderLibreoffice(PreviewBuilder):
                 'this builder requires libreoffice to be available'
             )
 
-
-    def build_jpeg_preview(
-            self,
-            file_path: str,
-            preview_name: str,
-            cache_path: str,
-            page_id: int,
-            extension: str = '.jpg',
-            size: ImgDims=None
-    ) -> None:
-
-        with open(file_path, 'rb') as odt:
-            if os.path.exists(
-                    '{path}{file_name}.pdf'.format(
-                        path=cache_path,
-                        file_name=preview_name
-                    )):
-                input_pdf_stream = open(
-                    '{path}.pdf'.format(
-                        path=cache_path + preview_name,
-                    ), 'rb')
-
-            else:
-                if self.cache_file_process_already_running(
-                                cache_path + preview_name):
-                    time.sleep(2)
-                    return self.build_jpeg_preview(
-                        file_path=file_path,
-                        preview_name=preview_name,
-                        cache_path=cache_path,
-                        extension=extension,
-                        page_id=page_id
-                    )
-
-                else:
-                    input_pdf_stream = convert_office_document_to_pdf(
-                        odt,
-                        os.path.splitext(file_path)[1],  # get the file extension
-                        cache_path,
-                        preview_name
-                    )
-
-            input_pdf = PdfFileReader(input_pdf_stream)
-            intermediate_pdf = PdfFileWriter()
-            intermediate_pdf.addPage(input_pdf.getPage(int(page_id)))
-
-            intermediate_pdf_stream = BytesIO()
-            intermediate_pdf.write(intermediate_pdf_stream)
-            intermediate_pdf_stream.seek(0, 0)
-            jpeg_stream = convert_pdf_to_jpeg(intermediate_pdf_stream, size)
-
-            jpeg_preview_path = '{path}{file_name}{extension}'.format(
-                path=cache_path,
-                file_name=preview_name,
-                extension=extension
-            )
-
-            with open(jpeg_preview_path, 'wb') as jpeg_output_stream:
-                buffer = jpeg_stream.read(1024)
-                while buffer:
-                    jpeg_output_stream.write(buffer)
-                    buffer = jpeg_stream.read(1024)
-
-    def get_page_number(self, file_path: str, preview_name: str,
-                        cache_path: str) -> int:
-
-        page_nb_file_path = cache_path + preview_name + '_page_nb'
-
-        if not os.path.exists(page_nb_file_path):
-            pdf_version_filepath = cache_path + preview_name + '.pdf'
-            if not os.path.exists(pdf_version_filepath):
-                self.build_pdf_preview(
-                    file_path=file_path,
-                    preview_name=preview_name,
-                    cache_path=cache_path
-                )
-
-            with open(page_nb_file_path, 'w') as page_nb_file_stream:
-                page_nb_file_stream.seek(0, 0)
-                with open(pdf_version_filepath, 'rb') as pdf_stream:
-                    pdf_reader = PdfFileReader(pdf_stream)
-                    page_nb_file_stream.write(str(pdf_reader.numPages))
-
-        with open(page_nb_file_path, 'r') as page_nb_stream:
-            page_nb = int(page_nb_stream.read())
-            return page_nb
-
-    def has_pdf_preview(self) -> bool:
-        """
-        Override and return True if your builder allow PDF preview
-        :return:
-        """
-        return True
-
-    def build_pdf_preview(
-            self,
-            file_path: str,
-            preview_name: str,
-            cache_path: str,
-            extension: str = '.pdf',
-            page_id: int = -1) -> None:
-
-        input_extension = os.path.splitext(file_path)[1]
-        intermediate_pdf_filename = preview_name.split('-page')[0] + '.pdf'
-        intermediate_pdf_file_path = os.path.join(
-            cache_path,
-            intermediate_pdf_filename
-        )
-
-        if not os.path.exists(intermediate_pdf_file_path):
-            if os.path.exists(intermediate_pdf_file_path + '_flag'):
-                # Wait 2 seconds, then retry
-                time.sleep(2)
-                return self.build_pdf_preview(
-                    file_path=file_path,
-                    preview_name=preview_name,
-                    cache_path=cache_path,
-                    extension=extension,
-                    page_id=page_id
-                )
-
-            with open(file_path, 'rb') as input_stream:
-
-                # first step is to convert full document to full pdf
-                convert_office_document_to_pdf(
-                    file_content=input_stream,
-                    input_extension=input_extension,
-                    cache_path=cache_path,
-                    output_filepath=intermediate_pdf_file_path
-                )
-
-        if page_id < 0:
-            return  # in this case, the intermediate file is the requested one
-
-        pdf_in = PdfFileReader(intermediate_pdf_file_path)
-        output_file_path = os.path.join(cache_path, '{}{}'.format(preview_name, extension))
-        pdf_out = PdfFileWriter()
-        pdf_out.addPage(pdf_in.getPage(page_id))
-
-        with open(output_file_path, 'wb') as output_file:
-            pdf_out.write(output_file)
-
-    def cache_file_process_already_running(self, file_name: str) -> bool:
-        if os.path.exists(file_name + '_flag'):
-            return True
-        else:
-            return False
-
-def create_flag_file(filepath: str) -> str:
-    """
-    Create a flag file in order to avoid concurrent build of same previews
-    :param filepath: file to protect
-    :return: flag file path
-    """
-    flag_file_path = '{}_flag'.format(filepath)
-    Path(flag_file_path).touch()
-    return flag_file_path
-
-
-def write_file_content(
+    def _convert_to_pdf(
+        self,
         file_content: typing.IO[bytes],
+        input_extension: str,  # example: '.dxf'
+        cache_path: str,
         output_filepath: str
-):
-    with open(output_filepath, 'wb') as temporary_file:
-        file_content.seek(0, 0)
-        buffer = file_content.read(1024)
-        while buffer:
-            temporary_file.write(buffer)
-            buffer = file_content.read(1024)
+    ) -> BytesIO:
+
+        return convert_office_document_to_pdf(
+            file_content, input_extension, cache_path, output_filepath
+        )
 
 
 def convert_office_document_to_pdf(
@@ -245,8 +82,13 @@ def convert_office_document_to_pdf(
         )
     # HACK - D.A. - 2018-05-31 - name is defined by libreoffice
     # according to input file name, for homogeneity we prefer to rename it
-    logging.debug('renaming output file {} to {}'.format(output_filepath+'.pdf', output_filepath))
-    os.rename(output_filepath+'.pdf', output_filepath)
+    # HACK-HACK - B.L - 2018-10-8 - if file is given without its extension
+    # in its name it won't have the double ".pdf"
+    if os.path.exists(output_filepath + '.pdf'):
+        logging.debug('renaming output file {} to {}'.format(
+            output_filepath + '.pdf', output_filepath)
+        )
+        os.rename(output_filepath + '.pdf', output_filepath)
 
     logging.debug('Removing flag file {}'.format(flag_file_path))
     os.remove(flag_file_path)
@@ -265,7 +107,7 @@ def convert_office_document_to_pdf(
 # HACK - D.A. - 2018-05-31
 # Code duplicated from https://raw.githubusercontent.com/LibreOffice/core/master/bin/get-bugzilla-attachments-by-mimetype
 LO_MIMETYPES = {
-# ODF
+    # ODF
     'application/vnd.oasis.opendocument.base': 'odb',
     'application/vnd.oasis.opendocument.database': 'odb',
     'application/vnd.oasis.opendocument.chart': 'odc',
@@ -287,7 +129,7 @@ LO_MIMETYPES = {
     'application/vnd.oasis.opendocument.text-template': 'ott',
     'application/vnd.oasis.opendocument.text-master-template': 'otm',
     'application/vnd.oasis.opendocument.text-web': 'oth',
-# OOo XML
+    # OOo XML
     'application/vnd.sun.xml.base': 'odb',
     'application/vnd.sun.xml.calc': 'sxc',
     'application/vnd.sun.xml.calc.template': 'stc',
@@ -301,7 +143,7 @@ LO_MIMETYPES = {
     'application/vnd.sun.xml.writer.global': 'sxg',
     'application/vnd.sun.xml.writer.template': 'stw',
     'application/vnd.sun.xml.writer.web': 'stw',
-# MSO
+    # MSO
     'application/rtf': 'rtf',
     'text/rtf': 'rtf',
     'application/msword': 'doc',
@@ -329,19 +171,19 @@ LO_MIMETYPES = {
     'application/vnd.visio2013': 'vsdx',
     'application/vnd.visio.xml': 'vdx',
     'application/x-mspublisher': 'pub',
-#WPS Office
+    # WPS Office
     'application/wps-office.doc': 'doc',
     'application/wps-office.docx': 'docx',
     'application/wps-office.xls': 'xls',
     'application/wps-office.xlsx': 'xlsx',
     'application/wps-office.ppt': 'ppt',
     'application/wps-office.pptx': 'pptx',
-# W3C
+    # W3C
     'application/xhtml+xml': 'xhtml',
     'application/mathml+xml': 'mml',
     'text/html': 'html',
     'application/docbook+xml': 'docbook',
-# misc
+    # misc
     'text/csv': 'csv',
     'text/spreadsheet': 'slk',
     'application/x-qpro': 'qpro',
@@ -352,8 +194,8 @@ LO_MIMETYPES = {
     'application/vnd.wordperfect': 'wpd',
     'application/wordperfect5.1': 'wpd',
     'application/vnd.ms-works': 'wps',
-    'application/clarisworks' : 'cwk',
-    'application/macwriteii' : 'mw',
+    'application/clarisworks': 'cwk',
+    'application/macwriteii': 'mw',
     'application/vnd.apple.keynote': 'key',
     'application/vnd.apple.numbers': 'numbers',
     'application/vnd.apple.pages': 'pages',
@@ -362,9 +204,9 @@ LO_MIMETYPES = {
     'application/x-iwork-pages-sffpages': 'pages',
     'application/x-hwp': 'hwp',
     'application/x-aportisdoc': 'pdb',
-    'application/prs.plucker' : 'pdb_plucker',
-    'application/vnd.palm' : 'pdb_palm',
-    'application/x-sony-bbeb' : 'lrf',
+    'application/prs.plucker': 'pdb_plucker',
+    'application/vnd.palm': 'pdb_palm',
+    'application/x-sony-bbeb': 'lrf',
     'application/x-pocket-word': 'psw',
     'application/x-t602': '602',
     'application/x-fictionbook+xml': 'fb2',
@@ -377,7 +219,7 @@ LO_MIMETYPES = {
     'application/x-starcalc': 'sdc',
     'application/x-stardraw': 'sdd',
     'application/x-starwriter': 'sdw',
-# relatively uncommon image mimetypes
+    # relatively uncommon image mimetypes
     'image/x-freehand': 'fh',
     'image/cgm': 'cgm',
     'image/tif': 'tiff',
@@ -392,10 +234,10 @@ LO_MIMETYPES = {
     'image/x-wmf': 'wmf',
     'image/x-pict': 'pict',
     'image/x-cmx': 'cmx',
-    # 'image/svg+xml': 'svg',  # nopep8 HACK - D.A. - 2018-07-05 Do not use libreoffice for SVG as inkscape is better
-    # 'image/bmp': 'bmp',
-    # 'image/x-ms-bmp': 'bmp',
-    # 'image/x-MS-bmp': 'bmp',
+    # 'image/svg+xml': 'svg',  # nopep8 HACK - D.A. - 2018-07-05 Do not use libreoffice for SVG as inkscape is better
+    # 'image/bmp': 'bmp',
+    # 'image/x-ms-bmp': 'bmp',
+    # 'image/x-MS-bmp': 'bmp',
     'image/x-wpg': 'wpg',
     'image/x-eps': 'eps',
     'image/x-met': 'met',
@@ -410,4 +252,3 @@ LO_MIMETYPES = {
     'image/x-xbitmap': 'xbm',
     'image/x-xpixmap': 'xpm',
 }
-
